# apps/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Rootly API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/api/workspaces/{workspaceId}/me/permissions": {
      "get": {
        "operationId": "getMyWorkspacePermissions",
        "summary": "Get My Workspace Permissions",
        "tags": [
          "Me"
        ],
        "description": "Get the authenticated user permissions in a workspace",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "workspaceId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "resource": {
                            "type": "string",
                            "enum": [
                              "workspace",
                              "folder",
                              "item",
                              "member",
                              "role"
                            ]
                          },
                          "action": {
                            "type": "string",
                            "enum": [
                              "create",
                              "read",
                              "update",
                              "delete",
                              "invite",
                              "all"
                            ]
                          }
                        },
                        "required": [
                          "resource",
                          "action"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "permissions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspaces/{workspaceId}/roles/{roleId}/permissions": {
      "get": {
        "operationId": "getRolePermissions",
        "summary": "Get Role Permissions",
        "tags": [
          "Roles"
        ],
        "description": "List all permissions for a role",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "workspaceId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "roleId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "permissions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "roleId": {
                            "type": "string"
                          },
                          "resource": {
                            "type": "string",
                            "enum": [
                              "workspace",
                              "folder",
                              "item",
                              "member",
                              "role"
                            ]
                          },
                          "action": {
                            "type": "string",
                            "enum": [
                              "read",
                              "create",
                              "update",
                              "delete",
                              "invite",
                              "all"
                            ]
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "roleId",
                          "resource",
                          "action",
                          "createdAt",
                          "updatedAt"
                        ],
                        "additionalProperties": false
                      }
                    }
                  },
                  "required": [
                    "permissions"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "setRolePermissions",
        "summary": "Set Role Permissions",
        "tags": [
          "Roles"
        ],
        "description": "Replace all permissions for a role",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "permissions": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "resource": {
                          "type": "string",
                          "enum": [
                            "workspace",
                            "folder",
                            "item",
                            "member",
                            "role"
                          ]
                        },
                        "action": {
                          "type": "string",
                          "enum": [
                            "read",
                            "create",
                            "update",
                            "delete",
                            "invite",
                            "all"
                          ]
                        }
                      },
                      "required": [
                        "resource",
                        "action"
                      ]
                    }
                  }
                },
                "required": [
                  "permissions"
                ]
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "workspaceId",
            "required": true
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "roleId",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    }
  }
}